+++ chaos-core/src/test/java/club/gclmit/chaos/core/pojo/Dog.java
@@ -210,7 +210,6 @@ package club.gclmit.chaos.core.pojo;
  * </p>
  *
  * @author gclm
- * @date 2020/5/2 3:37 下午
  * @since 1.8
  */
 public class Dog extends User {