+++ chaos-core/src/test/java/club/gclmit/chaos/core/pojo/GirlFriend.java
@@ -213,7 +213,6 @@ import java.util.Map;
  * </p>
  *
  * @author gclm
- * @date 2020/4/16 6:46 下午
  * @since 1.8
  */
 public class GirlFriend {