+++ chaos-core/src/test/java/club/gclmit/chaos/core/utils/DateUtilsTest.java
@@ -216,7 +216,6 @@ import java.time.ZoneOffset;
  * </p>
  *
  * @author gclm
- * @date 2020/3/27 10:35 上午
  * @since 1.8
  */
 public class DateUtilsTest {